added controller functions
