+++ router/orderRouter.js
@@ -1,110 +1,20 @@
 const express = require('express');
+// const moment = require('moment')
 const { authenticateUser } = require('../middleware/authenticate');
-const orderModel = require('../models/orderModel');
+// const orderModel = require('../models/orderModel');
+const orderController = require('../controller/orderController')
 
 const orderRouter = express.Router()
 
-orderRouter.post('/', authenticateUser,  async (req, res) => {
-    const body = req.body;
+orderRouter.post('/',authenticateUser, orderController.createNewOrder)
 
-    const total_price = body.items.reduce((prev, curr) => {
-        prev += curr.price
-        return prev
-    }, 0);
+orderRouter.get('/:orderId',authenticateUser, orderController.getOrderById)
 
-    const order = await orderModel.create({ 
-        items: body.items,
-        created_at: moment().toDate(),
-        total_price
-    })
+orderRouter.get('/', authenticateUser, orderController.getAllOrders)
+
+orderRouter.patch('/:id', authenticateUser, orderController.updateOrder)
+
+orderRouter.delete('/:id', authenticateUser, orderController.deleteOrder)
     
-    return res.json({ status: true, order })
-})
-
-orderRouter.get('/:orderId',authenticateUser, async (req, res) => {
-    const { orderId } = req.params;
-    const order = await orderModel.findById(orderId)
-    
-    if (!order) {
-        return res.status(404).json({ status: false, order: null })
-    }
-
-    return res.json({ status: true, order })
-})
-
-orderRouter.get('/', authenticateUser, async (req, res) => {
-    // check for query parameters
-    const { price, date, state } = req.query
-
-    // query by price
-    if(price){
-        let value;
-        if(price === 'asc'){
-            value = 1
-            const orders = await orderModel.find().sort( {total_price : value} )
-            return res.json({ status: true, orders })
-        }
-
-        if(price === 'desc'){
-            value = -1
-            const orders = await orderModel.find().sort( {total_price : value} )
-            return res.json({ status: true, orders })
-        }
-        // query by date
-    } else if(date){
-        let value;
-        if(price === 'asc'){
-            value = 1
-            const orders = await orderModel.find().sort( {created_at : value} )
-            return res.json({ status: true, orders })
-        }
-
-        if(price === 'desc'){
-            value = -1
-            const orders = await orderModel.find().sort( {created_at : value} )
-            return res.json({ status: true, orders })
-        }
-    } 
-    
-    // PAGINATION //
-    const page = req.query.page * 1
-    const limit = req.query.limit * 1
-    const skip = (page - 1) * limit
-
-    const orders = await orderModel.find().skip(skip).limit(limit)
-
-    return res.json({ status: true, orders })
-})
-
-orderRouter.patch('/:id', authenticateUser, async (req, res) => {
-    const { id } = req.params;
-    const { state } = req.body;
-
-    const order = await orderModel.findById(id)
-
-    if (!order) {
-        return res.status(404).json({ status: false, order: null })
-    }
-
-    if (state < order.state) {
-        return res.status(422).json({ status: false, order: null, message: 'Invalid operation' })
-    }
-
-    order.state = state;
-
-    await order.save()
-
-    return res.json({ status: true, order })
-})
-
-orderRouter.delete('/:id', authenticateUser, async (req, res) => {
-    const { id } = req.params;
-
-    const order = await orderModel.deleteOne({ _id: id})
-
-    return res.json({ status: true, order })
-})
-
-
 
 module.exports = orderRouter
